[백준] 23234 The World Responds - Swift
시간 제한	메모리 제한	제출	정답	맞힌 사람	정답 비율
1 초	1024 MB	930	860	829	92.729%
문제
In many introductory computer programming classes, the first program that students learn to write just prints “Hello, world!” It is used as a first assignment because it is a simple program that produces output. The program dates back to at least 1974, when Brian Kernighan used it as an example in a C programming tutorial. Its popularity has grown since then. Louisiana Tech University's ACM chapter has a Hello World project that has collected versions of this program in almost 200 different computer languages.

While “Hello, World!” is a nice and simple first program, it does have one drawback. The program says hello to the world, but the world does not respond! This one-sided conversation is starting to become awkward after more than 35 years. For this problem, write a program that returns greetings from the world to the program.

입력
There is no input for this problem.

출력
Print one line of output: The world says hello!

예제 입력 1  복사
예제 출력 1  복사
The world says hello!